Trim and clean up the Fast-LLM test suite
Apply the coarse + fine review findings:

- Delete test_actor_error_handling.py — every test reimplemented the
  actor error-handling logic inline and asserted on the copy, so it
  could not catch a regression in the real actor.
- Drop the generic-path topology matrix (TestWeightUpdateTP2,
  TestWeightUpdateMultiActor); equivalent multi-topology coverage lives
  on the Fast-LLM path in test_vllm1_fast_llm_broadcast.py.
- Move the EngineManager import out of conftest top level into the
  factory fixture so the CPU-only unit tests collect without vLLM/torch.
- Consolidate the duplicated stream_process_output / kill_process_tree
  helpers into server_weight_update_utils.py and drop the now-pointless
  injected stream_process_output_fn parameter.
- Remove dead helpers (force_kill_process, check_pattern_detected),
  unused fixtures/params/imports, and the dead timeout parameter.
- Minor typing and style fixes (builtin generics, strict zip, f-string
  and implicit-Optional cleanups).
